+++ .changeset/spotty-donkeys-admire.md
@@ -1,0 +1,5 @@
+---
+"create-t3-app": minor
+---
+
+feat: add trpc rsc prefetching
